requirements.txt
cffi==1.7.0
Flask==0.11.1
Flask-Login==0.3.2
Flask-Bcrypt==0.7.1
pymongo==3.3.0
requests==2.10.0
APScheduler==3.2.0
twython==3.4.0
nltk==3.2.1
slacker==0.9.28
pyxero==0.8.0